(#247) Enable packageHashValidation feature
With Chocolatey 2.3.0 we introduced a feature
which will validate the checksum of a downloaded
nupkg with the SHA512 checksum that the repository
reports.

This increases confidence that the nupkg you are
installing is in fact the nupkg you expect.

This change enables the feature on the server as
it is being setup, and adds the command to turn on
the feature to the ClientSetup script.